requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
accelerate>=0.24.0

# Ollama integration
ollama>=0.1.7

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Evaluation and monitoring
evaluate>=0.4.0
wandb>=0.15.0

# UI and interface
gradio>=3.50.0
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.25.0

# RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pypdf>=3.17.0
docx2txt>=0.8
beautifulsoup4>=4.12.0
requests>=2.31.0
